% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm_plot.R
\name{btm_topic_model}
\alias{btm_topic_model}
\title{BTM Topic Modeling}
\usage{
btm_topic_model(df, k = 10, iter = 1000)
}
\arguments{
\item{df}{Annotated dataframe containing the text data.}

\item{k}{Number of topics.}

\item{iter}{Number of iterations for the Gibbs sampling algorithm.}
}
\value{
A BTM model.
}
\description{
This function applies the Biterm Topic Model (BTM) to a given annotated dataframe.
}
\examples{
data <- data.frame(doc_id = c(1,2,3), upos = c("NOUN", "PROPN", "NOUN"), lemma = c("noun1", "noun2", "noun3"))
model <- btm_topic_model(data, k = 5, iter = 1000)
}
