% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{btm_topic_model_cooccurrence}
\alias{btm_topic_model_cooccurrence}
\title{BTM Topic Modeling with Cooccurrence}
\usage{
btm_topic_model_cooccurrence(df, k = 10, iter = 2000)
}
\arguments{
\item{df}{Annotated dataframe containing the text data.}

\item{k}{Number of topics.}

\item{iter}{Number of iterations for the Gibbs sampling algorithm.}
}
\value{
A BTM model.
}
\description{
This function applies the Biterm Topic Model (BTM) to a given annotated dataframe,
using cooccurrence of terms to build the model.
}
\examples{
data <- data.frame(doc_id = c(1,2,3), xpos = c("NN", "VB", "NN"), token = c("token1", "token2", "token3"))
model <- btm_topic_model_cooccurrence(data, k = 5, iter = 1000)
library(stopwords)
# Read data from CSV
df <- read.csv("filtered_data_18_2023.csv")
df <- tail(df, 600)
df$Description <- iconv(df$Description, from = "ISO-8859-1", to = "UTF-8")
# Load the udpipe model
ud_model <- udpipe_load_model("english-ewt-ud-2.5-191206.udpipe")

# Annotate the text data
u_df <- text_annotation(df, "Description", ud_model)
# Run Biterm Topic Model with cooccurrence
btm_model_cooccurrence <- btm_topic_model_cooccurrence(u_df, k = 10)

# Plot the topic model
plot(btm_model_cooccurrence)


}
