% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cves_harvesting.R
\name{cves_harvesting}
\alias{cves_harvesting}
\title{Download CSV from a URL}
\usage{
cves_harvesting(
  url = "https://cve.mitre.org/data/downloads/allitems.csv",
  file_name = "allitems.csv"
)
}
\arguments{
\item{url}{The URL of the CSV file. Default is "https://cve.mitre.org/data/downloads/allitems.csv".}

\item{file_name}{The name of the file where the CSV content will be stored. Default is "allitems.csv".}
}
\value{
A boolean indicating if the file was successfully downloaded.
}
\description{
This function downloads a CSV file from a given URL and stores it into a local file.
}
\examples{
cves_harvesting()
cves_harvesting("https://cve.mitre.org/data/downloads/allitems.csv", "allitems.csv")

cves_harvesting()
}
