% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter Data}
\usage{
filter_data(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A filtered data frame.
}
\description{
Filters a data frame based on specific conditions.
}
