% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cves_harvesting.R
\name{read_data_CVEs}
\alias{read_data_CVEs}
\title{Read Data from CVEs CSV File}
\usage{
read_data_CVEs(file_name)
}
\arguments{
\item{file_name}{The name of the file to read.}
}
\value{
A dataframe with data from the CSV file.
}
\description{
This function reads a CSV file, skipping the first few rows to retrieve the headers and data separately.
It then assigns the header names to the data and returns a dataframe.
}
\examples{
read_data_CVEs("allitems.csv")

cves_harvesting("https://raw.githubusercontent.com/Cdk29/cveR/main/tests/testthat/allitems_test.csv", "test_download_from_github.csv")
CVEs_df <- read_data_CVEs("test_download_from_github.csv")
head(CVEs_df, 2)
colnames(CVEs_df)
}
