% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{text_annotation}
\alias{text_annotation}
\title{Text Annotation}
\usage{
text_annotation(df, text_col, ud_model)
}
\arguments{
\item{df}{Dataframe containing the text data.}

\item{text_col}{Name of the column in df that contains the text data.}

\item{ud_model}{The udpipe model to use for POS tagging.}
}
\value{
A data frame with annotations from the udpipe model.
}
\description{
This function applies the udpipe model to a given dataframe and a specified text column for POS tagging.
}
\examples{
data <- data.frame(description = c("text1", "text2", "text3"))
ud_model <- udpipe_load_model("english-ud-2.0-170801.udpipe")
annotated_data <- text_annotation(data, "description", ud_model)
u_df <- text_annotation(test_df, "Description", ud_model)
head(u_df)
}
