% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm_plot.R
\name{topic_model_BTM_plot_generation}
\alias{topic_model_BTM_plot_generation}
\title{BTM Topic Model Plot Generation}
\usage{
topic_model_BTM_plot_generation(df, ud_model_file)
}
\arguments{
\item{df}{A dataframe containing the text data.}

\item{ud_model_file}{A string specifying the location of the udpipe model file.}
}
\value{
Generates a plot of the Biterm Topic Model.
}
\description{
This function takes a dataframe and a udpipe model file, applies text annotation and the Biterm Topic Model (BTM), and then generates a plot of the BTM.
}
\examples{
df <- read.csv("filtered_data_18_2023.csv")
ud_model_file <- "english-ewt-ud-2.5-191206.udpipe"
topic_model_BTM_plot_generation(df, ud_model_file)
}
