% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{update_cve}
\alias{update_cve}
\title{Update the CVE dataframe}
\usage{
update_cve(new_cves, archived_cve)
}
\arguments{
\item{new_cves}{Dataframe The dataframe with new CVEs.}

\item{archived_cve}{Dataframe The dataframe with archived CVEs.}
}
\value{
Dataframe The updated archived_cve dataframe with new CVEs added.
}
\description{
Add the CVEs present in new_cves dataframe that are not yet present in archived_cve dataframe.
The new CVEs are added with the Project as 'new_cves' and Archived status as 'No'.
}
\examples{
updated_cve <- update_cve(new_cves, archived_cve)
library(dplyr)
library(tidyverse)
archived_cve <- update_cve(new_cves, archived_cve)
}
